requirements.txt
Django>=3.0.7
psycopg2-binary>=2.8.5
djangorestframework>=3.10.3
requests==2.25.1
sqlalchemy==1.4.1
